+++ weblate/trans/tests/test_hooks.py
@@ -681,6 +681,77 @@ AZURE_PAYLOAD = r'''
 }
 '''
 
+GITEA_PAYLOAD = '''
+{
+  "secret": "3gEsCfjlV2ugRwgpU#w1*WaW*wa4NXgGmpCfkbG3",
+  "ref": "refs/heads/master",
+  "before": "28e1879d029cb852e4844d9c718537df08844e03",
+  "after": "bffeb74224043ba2feb48d137756c8a9331c449a",
+  "compare_url": "http://localhost:3000/gitea/webhooks/compare/28e187...bffeb7422",
+  "commits": [
+    {
+      "id": "bffeb74224043ba2feb48d137756c8a9331c449a",
+      "message": "Webhooks Yay!",
+      "url": "http://localhost:3000/gitea/webhooks/commit/bffeb74224043ba2feb4",
+      "author": {
+        "name": "Gitea",
+        "email": "someone@gitea.io",
+        "username": "gitea"
+      },
+      "committer": {
+        "name": "Gitea",
+        "email": "someone@gitea.io",
+        "username": "gitea"
+      },
+      "timestamp": "2017-03-13T13:52:11-04:00"
+    }
+  ],
+  "repository": {
+    "id": 140,
+    "owner": {
+      "id": 1,
+      "login": "gitea",
+      "full_name": "Gitea",
+      "email": "someone@gitea.io",
+      "avatar_url": "https://localhost:3000/avatars/1",
+      "username": "gitea"
+    },
+    "name": "webhooks",
+    "full_name": "gitea/webhooks",
+    "description": "",
+    "private": false,
+    "fork": false,
+    "html_url": "http://localhost:3000/gitea/webhooks",
+    "ssh_url": "ssh://gitea@localhost:2222/gitea/webhooks.git",
+    "clone_url": "http://localhost:3000/gitea/webhooks.git",
+    "website": "",
+    "stars_count": 0,
+    "forks_count": 1,
+    "watchers_count": 1,
+    "open_issues_count": 7,
+    "default_branch": "master",
+    "created_at": "2017-02-26T04:29:06-05:00",
+    "updated_at": "2017-03-13T13:51:58-04:00"
+  },
+  "pusher": {
+    "id": 1,
+    "login": "gitea",
+    "full_name": "Gitea",
+    "email": "someone@gitea.io",
+    "avatar_url": "https://localhost:3000/avatars/1",
+    "username": "gitea"
+  },
+  "sender": {
+    "id": 1,
+    "login": "gitea",
+    "full_name": "Gitea",
+    "email": "someone@gitea.io",
+    "avatar_url": "https://localhost:3000/avatars/1",
+    "username": "gitea"
+  }
+}
+'''
+
 
 class HooksViewTest(ViewTestCase):
     @override_settings(ENABLE_HOOKS=True)
@@ -716,6 +787,17 @@ class HooksViewTest(ViewTestCase):
         response = self.client.post(
             reverse('webhook', kwargs={'service': 'github'}),
             {'payload': GITHUB_NEW_PAYLOAD}
+        )
+        self.assertContains(response, 'Update triggered')
+
+    @override_settings(ENABLE_HOOKS=True)
+    def test_hook_gitea(self):
+        # Adjust matching repo
+        self.component.repo = 'http://localhost:3000/gitea/webhooks.git'
+        self.component.save()
+        response = self.client.post(
+            reverse('webhook', kwargs={'service': 'gitea'}),
+            {'payload': GITEA_PAYLOAD}
         )
         self.assertContains(response, 'Update triggered')